\name{ggarea}
\alias{ggarea}
\title{Area grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{ggarea(plot = .PLOT, aesthetics=list(), ..., data=plot$data)}
\arguments{
\item{plot}{x positions}
\item{aesthetics}{y positions}
\item{...}{id variable used to separate observations into different areas}
\item{data}{colour}
\item{}{pattern}
\item{}{...}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
