\name{ggorder}
\alias{ggorder}
\title{Order plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
A plot to investigate the order in which observations were recorded.
}
\usage{ggorder(data, scale="rank")}
\arguments{
\item{data}{}
\item{scale}{}
}

\details{ar
Need ggobi version as well that creates edge between consecutive observations (and adds row number to dataset)}

\examples{}
\keyword{hplot}
