\name{ggplot.options}
\alias{ggplot.options}
\title{Set ggplot options}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set global options for ggplot.
}
\usage{ggplot.options(...)}
\arguments{
\item{...}{list of options to get/set}
}

\details{\item[grid.col] colour of background grid lines, ("white")
\item[grid.fill] colour of background, ("grey90")}

\examples{}
\keyword{manip}
