\name{map}
\alias{map}
\alias{map.scale}
\alias{map.categorical}
\alias{map.continuous}
\alias{map.gradient}
\alias{map.ps_double}
\alias{map.ps_map}
\alias{map.scale_cont_colour}
\alias{map.manual}
\title{Map}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map raw data to aesthetics using the provided scale
}
\usage{map(scale, data, ...)}
\arguments{
\item{scale}{}
\item{data}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
