\name{print.ggplot}
\alias{print.ggplot}
\title{Print ggplot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print generic for ggplot.  Plot on current graphics device.
}
\usage{print.ggplot(x, ...)}
\arguments{
\item{x}{plot to display}
\item{...}{other arguments passed on to \code{\link{ggplot_plot}}}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
